Improve scroll to bottom behavious
